Adding an additional interface for Atm
