+++ hw06-atm/src/main/java/org/mlaptev/otus/atm/AtmCaretaker.java
@@ -2,15 +2,15 @@ package org.mlaptev.otus.atm;
 
 import org.mlaptev.otus.exceptions.AtmException;
 
-public class AtmCaretaker {
+class AtmCaretaker {
 
   private Object obj;
 
-  public void save(Atm atm) {
+  void save(Atm atm) {
     this.obj = atm.save();
   }
 
-  public void undo(Atm atm) throws AtmException {
+  void undo(Atm atm) throws AtmException {
     atm.undo(this.obj);
   }
 }
